chore: Move a11y violation tests to Playwright (#30293)
