+++ apps/meteor/package.json
@@ -63,6 +63,7 @@
 		"email": "support@rocket.chat"
 	},
 	"devDependencies": {
+		"@axe-core/playwright": "^4.7.3",
 		"@babel/core": "~7.22.9",
 		"@babel/eslint-parser": "~7.22.9",
 		"@babel/plugin-proposal-nullish-coalescing-operator": "~7.18.6",
@@ -115,7 +116,6 @@
 		"@types/he": "^1.1.2",
 		"@types/i18next-sprintf-postprocessor": "^0.2.0",
 		"@types/imap": "^0.8.37",
-		"@types/jest-axe": "^3.5.5",
 		"@types/jsdom": "^16.2.15",
 		"@types/jsdom-global": "^3.0.4",
 		"@types/jsrsasign": "^10.5.8",
@@ -183,7 +183,6 @@
 		"fast-glob": "^3.2.12",
 		"i18next": "~23.4.5",
 		"jest": "~29.6.1",
-		"jest-axe": "^8.0.0",
 		"jsdom-global": "^3.0.2",
 		"mocha": "^9.2.2",
 		"nyc": "^15.1.0",
